removed obsolete code
